test: 메뉴 도메인 이름 테스트 코드 추가
