+++ README.md
@@ -99,10 +99,10 @@ docker compose -p kitchenpos up -d
 ### 상품
 
 | 한글명 | 영문명        | 설명                 |
-| --- |------------|--------------------|
+|----|------------|--------------------|
 | 상품 | product    | 메뉴를 관리하는 기준이 되는 데이터 |
 | 이름 | clean cleanName | 음식을 상상하게 만드는 중요한 요소 |
-| 이름 | price      | 음식의 가격             |
+| 가격 | price      | 음식의 가격             |
 
 ### 메뉴
