+++ src/test/java/kitchenpos/menus/domain/MenuGroupTest.java
@@ -1,11 +1,8 @@
 package kitchenpos.menus.domain;
 
-
-import kitchenpos.menus.tobe.domain.CleanName;
 import kitchenpos.menus.tobe.domain.MenuGroup;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
@@ -30,27 +27,5 @@ public class MenuGroupTest {
         Assertions.assertThatThrownBy(
                 () -> new MenuGroup(UUID.randomUUID(), 이름)
         ).isInstanceOf(IllegalArgumentException.class);
-    }
-
-    @Nested
-    @DisplayName("메뉴 그룹 이름 테스트")
-    class NameTest {
-        @Test
-        @DisplayName("메뉴 그룹의 이름을 생성한다.")
-        void create() {
-            String 이름 = "후라이드";
-            CleanName 메뉴_그룹_이름 = new CleanName(이름);
-
-            Assertions.assertThat(메뉴_그룹_이름.getName()).isEqualTo(이름);
-        }
-
-        @NullAndEmptySource
-        @ParameterizedTest
-        @DisplayName("메뉴 그룹 이름은 공백일 수 없다.")
-        void create_exception_empty(String 이름) {
-            Assertions.assertThatThrownBy(
-                    () -> new CleanName(이름)
-            ).isInstanceOf(IllegalArgumentException.class);
-        }
     }
 }
